LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-tight-padding-expected.txt

Testing MacOSInlineMediaControls with tight padding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.leftContainer.leftMargin is 12
PASS mediaControls.leftContainer.rightMargin  is 12
PASS mediaControls.leftContainer.buttonMargin is 12
PASS mediaControls.rightContainer.leftMargin  is 12
PASS mediaControls.rightContainer.rightMargin is 12
PASS mediaControls.rightContainer.buttonMargin  is 12

PASS successfullyParsed is true

TEST COMPLETE
